Add "->?" as the safe indexing variant of "->"
This was earlier covered by "?->", but this is inconsistent with the
"[?" index variant and the future "(?" program indexing variant.

Usage of "?->" will start emitting a deprication warning is some
future major version of Pike.

Manual backport from master: 19582df85018fdb488de25c340bfa1efc44becd0,
because the conflict was too big to bother with cherry-pick.