Optimizer: Make Coverity happier
Coverity does not like us using the result of arguments(1) ($1a) as
array index, as arguments() could in theory return -1. Guarding against
that seems not possible in peep.in rules, so we now instead just call to
a function added to peep.c that performs the check in a (hopefully)
Coverity-approved way [CID 1455693].

This is probably not a very elegant solution, however I assume guarding
in the form of [$1a >= 0 && ...] would not make Coverity happy since it
probably cannot infer that arguments(1) returns the same value both
times. But I really do not know the extent of Coverity's magic here.