Backend: Fix "Free call_out in call_out hash table"
remove_call_out() removed the call out from the heap but did not
unlink in from the hashtables.  Instead it was unlinked by the EXIT
callback, even if pos was -1.  However, backend_verify_call_outs()
does not like call outs which are in the hashtable while pos is -1.
So change remove_call_out() to do the same cleanup that
backend_do_call_outs() does.