Backend: Improved fix for busy-wait on short call_outs (< 1ms) and poll.
The conversion from µs to ms now rounds up (instead of down), and the minimum
wait of 2ms is removed. This makes it possible to (once again) schedule call_outs
to be called immediately (instead of in 2ms).

Thanks to Tomasz Jamroszczak <tjamroszczak@opera.com> for the report,
and to Tobi for the analysis.

Fixes [LysLysKOM 23729673].