SSL.File: Signal errors during handshake via close callback
We relatively recently stopped claiming that a connection was
peer_closed when in fact we gave up on the SSL connection with a fatal
alert. This caused us to signal some errors via the write callback
instead. This restores the old signalling behaviour by checking for
local_fatal state in some places we only checked for peer_closed before.