Compiler: Get rid of the top_statement_label_dummy sentinel.
This label sometimes got cleanup state attached to it. As it
was shared between multiple compilations that was a BAD idea,
and could cause some very strange runtime errors.

Fixes errors such as "Catch context doesn't belong to this frame."
(which was caused by do_escape_catch in the sentinel cleanups).

docode() and do_code_block() now allocate a new statement_label.