Process [NT]: Fix process creation.
The switch to using STARTUPINFOEXW caused CreateProcessW() to
fail with ERROR_INVALID_PARAMETER (87) as soon as the
EXTENDED_STARTUPINFO_PRESENT flag was set (even with an empty
attribute list). This seems to have been caused by GetStartupInfoW()
having filled in several reserved fields of the StartupInfo.
As none of the information that it filled in seems relevant,
the call has now been removed. Starting of processes on NT
should now work again.