Stdio.File: Added tcdrain() and tcsetsize().
It is now possible to set the window size for a pty or other terminal.

Also adds documentation for them as well as tcflush() and tcsendbreak(),
and moves it and the documentation for tc{get,set}attr() from Stdio.File
to Stdio.Fd.

Also some fixes to make the code work on systems with <sys/termios.h>
but no <termios.h>, and preparation to make some of it work on NT.

Fixes some of #10013.