SSL.Context: Context commits to tickets via offers_tickets()
In certain situations, i.e. after receiving an empty Session Ticket
extension (client supports, but does not yet have a ticket), by
including the Session Ticket extension in the server hello commits us to
providing a ticket later.

As SSL.Context would simply return 0 in encode_ticket() when use_cache
was zero (not the default), we used to break the protocol on such
occasions. Instead, we now use SSL.Context()->offers_tickets() to signal
whether Session Tickets should be supported on the connection (and
whether the Context commits to encoding a session as ticket later on).