Build [Configure]: Blacklist getpagesize on macOS
getpagesize() was removed from POSIX.1-2001.  macOS therefore removes
the prototype from <unistd.h> if _POSIX_C_SOURCE is 200112 or higher.
However, the function still exists for backwards compat, which leads
to the configure check finding it (since it fakes a prototype).
Prevent the configure check from running and finding the prototype-less
function so that we don't get a compilation error trying to call it.
(The correct way of getting the page size in POSIX is sysconf(_SC_PAGESIZE).)