Image.PSD: Fix size check
The additional size checks introduced for the color table size
unfortunately did not cover all cases in which the color table was used,
and instead did cover cases in which the color table would not be used
for decoding the PSD image, leading to a null dereference.

Thanks to Cezary Cerekwicki <ccerekwicki@opera.com> for the report.