Testsuite: Relax upper bound of gethrtime() test
On macos Big Sur, the sleep(2.0) call can overshoot the target by
0.12-0.15 s if the system is otherwise idle (this is presumably
related to the "napping" feature).  This is not really an error since
sleep only guarantees that _at least_ the requested time will elapse.

Adjust the upper bound of the accepted interval to be T+0.2, but
keep the lower bound at T-0.1.