Concurrent: Fix Future()->timeout()
call_callback() may use call_out(cb, 0, ...) to call a callback. This
means that if we call multiple callbacks wie call_callback(), they mæy
be executed in arbitrary order. The implementation of
Future()->timeout() was subject to races in this regard, leading to the
Future returned by ->timeout() to resolve with 0 instead of the result
or instead of rejecting with the same reason as the original
Promise/Future was rejected before the timeout triggered.

Using less abstractions, we save on such races, Future objects, and even
lambdas while getting arguably eaiser to read code.