Compiler: Support F_CLEAR_LOCAL in the parse tree.
This node is equivalent to F_ASSIGN with 0, but only valid for F_LOCAL
variables, and only when the variable will NOT be used afterwards.

The intent is to simplify avoiding of errors from the type system for
the case where the variable has a type that does not include zero.