Compiler [Typechecker]: Fixed reference handling in store_marker()
In cases where the value passed to store_marker() was NULL but the
marker was present in the mapping, we used to overwrite the marker with
itself. By doing so, assign_svalue() (as used by low_mapping_insert())
would free the same value we were inserting. This is never great but has
especially noticeable consequences when the value only had a single
reference.

Instead, we now simply do not re-insert the value in these cases.