Configure: Fix failure of FSEvents to configure with --disable-binary
FSEvents uses AC_CHECK_DECLS in its configure script.  AC_CHECK_DECLS
has an internal test where it tries to figure out what flag to pass
to the compiler to make compilation of a certain piece of code _fail_.
But nobinary_dummy "succeeds" at compiling anything, which makes it
impossible for this configure test to succeed.  Work around the issue
by pre-seeding the cache with the desired outcome for the internal test.

Also, fix the incorrect test guarding the assignment of CC to
$BINDIR/nobinary_dummy.