__builtin.Nettle.Hash: Fix some warnings.
Fixes fall-out from the stricter type for predef::`*().

Stricter types for the block_size argument in some functions,
where it now has the type int(1..)|void (previously int|void)).

Also uses the new default argument value syntax is a few places.
This means that a few places that previously accepted 0 as a
placeholder for the default value now require UNDEFINED.