Compat [8.0]: Use symbols directly from _Debug.
Change (indirect) references to symbols from the Debug module
to direct references to same in the _Debug module.

The Debug module contains code that pulls in lots of other
modules (among other things Stdio.Buffer). This caused the
early bootstrap code to load stuff in an unsupported order
when run -V8.0, causing the Stdio module to fail to compile.

Fixes support for -V8.0 (and earlier).