EFUNs: Reduce strictness of the aggregate_mapping() result.
The previous type was correct, but generated huge types that
caused even larger performance issues in the type checker.
The resulting type should now be on par with the one generated
by the old hand-coded type-deriver for aggregate_mapping().

Fixes issues deriving the type for the timezone_expert_tree
mapping in Calendar.TZnames.