Compiler: Retain flags for local vars used in scope
We require the LOCAL_VAR_USED_IN_SCOPE flag during optimization.
Previously, during optimization, we checked uninitialized memory when
decoding dumped modules to decide if certain optimizations apply. Now we
initialize the local variable information to zero and dump the flags to
local variables when required, so that they can be re-applied when
decoding.