Compiler: Fix support for cast in argument default value expressions.
The new_arg_name production uses full_type, which means that there
is still a type on the type stack when the optional_default_value
production parses its expression. This confused cast et al as they
expect the type stack to be empty.

Fixes fatal "pop_unfinished_type(): Unexpected len: 2" that was
triggered by the cast to int in the following code in Standards.X509:

  string make_selfsigned_certificate(Crypto.Sign.State c, int ttl,
  [...]
				     int serial =
				     (int)Gmp.mpz(Standards.UUID.
						  make_version1(-1)->encode(),
						  256))
  [...]