Git: Begin migration to using only spaces for indentation.
There are apparently developers that insist on using editors that
can not distinguish between tab size and indentation depth. Those
editors get confused by code (like Pike) that use a mix of tabs
and spaces for indentation and where the tab size (8) differs from
the base indentation depth (typically 2).

This commit sets the git attribute "space_indent" on all files,
defines it to use the git filter "space-indent" and adds a list
of exceptions (that also get the attribute "mixed_indent").

The space-indent filter is typically defined with:
$ git config --global filter.space-indent.clean 'unexpand --first-only'
$ git config --global filter.space-indent.smudge 'cat'

githelper has been updated to know about the attributes.