Compiler: Clean up copy_node().
Now that the node hash is no more, there's no reason to perform
actual copying of F_LOCAL and F_TRAMPOLINE nodes. Note also that
the copying of F_LOCAL_INDIRECT nodes was a bad idea, since then
the copy would not get bound to the same F_LOCAL as the original.

Fixes "Indexing the NULL value" when eg using the dynamic dot
operator on an unbound local variable.

Fixes multiple testsuite failures.