Build [NT]: Simplified STATIC_ASSUME() macro.
The UNREACHABLE() macro is always available, so there's no need
for a special case when __builtin_unreachable() is not. This
causes the argument to STATIC_ASSUME() to always be seen by the
compiler.

Fixes lots of warnings about __B (from the ASSUME_UNCHANGED()) macro
in buffer.h being unused on eg NT.