sprsh: Improved handling of EOF on stdin.
The sprsh-side now sends an out-of-band '\4' when stdin
reaches EOF. Previously it killed itself with SIGINT,
expecting that the signal handler would send an out-of-
band 'x'. In ancient versions of nt-tools EOF on stdin
was signalled by an uni-directional close of the socket.
This was apparently disabled due to issues with some
virtualization environments.

Ending a Hilfe-session by pressing Ctrl-D now works again.