Cpp: Fix #if constant(.Foo)
The call to handle_import() when resolving module identifiers starting
with '.' should use the special relative path "" to indicate the current
directory, just like it is done for 'import .Foo'.  The reason that
handle_import(".") is wrong is that it relies on the intepretation of
an entry "." in a directory to be a reference back to the directory
itself, which is operating environment specific.