Compiler: Clean up the expression grammar.
The base expression grammar now mirrors that of the C-standard.

This makes it easier to understand what the various expression
productions mean.

Also adds the unary '+' operator (noop) for completion.

NOTE: The power operator (`**) now has the same precedence
as in D (albeit with a different token). This means that it
binds slightly harder than before. It now binds harder than
the preincrement and predecrement operators.