Compiler [AST]: Add set_node_name().
This function sets the (previously unused) name field for a
node, and is intended to be a description of the parsetree
rooted in the node useable for compiler diagnostics.

Eg: The name for an F_CONSTANT node could be set to
the symbol used to lookup the constant.

This is in preparation for some further grammar adjustments.