Testsuite [mktime]: Check correct operation for years < 1900
The workaround for mktime on AIX and NT not supporting negative
time_t:s is based on the principle that "The calendar repeats every 28
years".  While this is true for the Julian calendar, this is
unfortunately not the case for the Gregorian calendar, which makes
year 0 modulo 100 not a leap year, unless it is also year 0 modulo
400.  As a result, the workaround yields incorrect results for any
date before the 1st of March 1900.