decode_value: Fix improved placeholder cleanup handling
The change of the local variable `placeholder' to become a sort of
local cache for `c->placeholder' broke the test `if(placeholder)',
which expected `placeholder' to be NULL in the case where a
placeholder exists but the decoding has become delayed (in this case
the placeholder will be carried over to the next pass as
`data->support_compilation->placeholder', and initialization should be
deferred to that pass).

Fix this by introducing another variable which explicitly carries the
information on whether or not the cached placeholder should be
initialized.