Compiler: Fix NULL dereference when a placeholder obj but no prog is provided
It is possible for c->target to be NULL while c->placeholder is non-NULL.
Therefore, use c->p as the target progam instead of c->target when checking
whether to run the placeholder initializers.  If a target program _was_
specified, c->p is guaranteed to be equal to c->target at that point.