Testsuite [7.8::SSL.sslfile]: Fixed issue with buffering in socketpair().
There is a race between the data being sent by the client arriving
at the server and the out-of-band signalling that the client is done.

The race typically caused truncation of data at the server due to
the end of transmission arriving before all data. This was mostly
noticed on platforms where socketpair_ultra() is used (ie NT).