Modules: Protect against creating static strings from dynamic modules.
Static strings created from dynamic modules may lead to
access of freed memory when the module has been unloaded.

The protection is done by redirecting those calls to calls
creating normal strings.

The issue is currently observed with the Java module, which may
allocate a static string "CLASSPATH" and do a getenv() on it.
When the module is unloaded the string may stay behind as it
could be referenced from the env mapping. Then the next time a
string is created with the same hash as "CLASSPATH" an invalid
pointer will be followed.