Compiler: Keep scope-related flags when re-adding prototypes.
This makes it possible to re-add a prototype for a local function
that uses lexically scoped variables in the last pass of the compiler.
Previously doing this caused the flags to be (temporarily) lost
causing some uses of the function to lose the trampoline property.

Preparation for reducing the amount of code-duplication
in the compiler somewhat.