Compiler: Initialize gc_marker in fake objects as well
low_start_new_program() creates a fake object that GC_ALLOC() is never
applied on, presumably so that it is not included in some counters
accounting for the number of objects and allocations. However, this left
the gc_marker of the fake object uninitialized, leading to invalid
memory access and thus undefined behaviour.

Initializing the marker in it is the most straight-forward fix, however
it seems that we might want to initialize or GC_ALLOC() all objects in
alloc_object(), and then just make sure they're also all GC_FREE()'d
later, so that the accounting stays roughly correct.