Compiler: Short-circuit safe indexing operators as per similar languages.
The safe-indexing operators (->?, [?] and (?)) now short-circuit
the expression when the operand is false. Previously later unsafe
operators were still applied. Eg:

  mapping foo = ([]);
  mixed q = foo->bar->?baz->foobar;

Previously the above caused a Indexing the NULL value with "foobar"
error. Now it instead results in q being assigned the value UNDEFINED
(ie the value of foo->bar). Note that the error will still be thrown
if the safe indexing operation itself succeeded. Eg:

  mapping q = foo->?bar->baz->foobar;

Will cause a Indexing the NULL value with "baz" error.

Fixes most of #10162.