Compiler: Fix some bugs in make_default_lvalue().
References to F_LOCAL_INDIRECT variables were handled incorrectly
(the stack offset was used where a local variable number was expected).

The argument and return types for the generated lambda was taken from
the type of the default value (rather than from the type of the variable).

The argument variable was not bound at declaration.