Runtime: Fix bug in stack_unlink() in combo with PROGRAM_DESTRUCT_IMMEDIATE.
If the svalue at the destination offset triggered a destroy() of
an object with a destroy() callback, the value to keep got zapped.
This was due to stack_unlink() decreasing the stack pointer while
assuming that the now floating entry would be unmodified after
freeing the destination svalue.

Fixes some obscure failures and crashes.