Tools.Shoot: Don't use unbounded memory in ArrayAdding benchmarks
The global variable variants of the ArrayAdding benchmarks kept
adding elements to the same array for the duration of the test
(3 seconds by default) without ever clearing it.  This caused
the array to grow very large on fast machines.  Change the tests
to clear the array after every 10000 additions, just like the local
variable based benchmark does.  This has just a small effect on the
result (about 1%) but a large one on memory use.