decode_value: Extend skip_values to rest of decode
This is because subsequent subclasses may contain references to
unresolved constants in the form of "again" tags, which do not trigger
a supporter dependency for that class.  But we know that the whole
encoded value will be parsed again, so we can safely ignore unresolved
constants for the rest of the current parse.