Operators: Fix range check in float to int conversion
If FLOAT_TYPE and and INT_TYPE have the same size (e.g. 64 bit)
then MAX_INT_TYPE will not be exactly representable as a FLOAT_TYPE
due to the mantissa overflowing.  This in turn will make the comparison
inexact.  Fix this by comparing with MAX_INT_TYPE+1, which is exactly
representable, instead.