Core [NT]: Inhibit warning about unused label from MSVC mk II.
 #pragma warning(disable:4102) was apparently not sufficient.

This may have been due to the scope for the pragma being on a function
context rather than on a basic block context. The pragma has now been
moved to cover the entirety of debug_check_rb_tree(), and the alternative
work-around has also been enabled. Either ought to fix the warning.