Build: Fix compatibility with GCC 5.2.
The 'malloc' attribute implies that the function returns a unique
pointer, and thus any assignments to the allocated pointer may be
removed by the optimizer if the pointer is not passed along.

This fixes issues with peep.c when compiled with gcc 5.2.1.

Thanks to Richard Biener for the triage.
Cf https://gcc.gnu.org/bugzilla/show_bug.cgi?id=123001

Fixes #10209.